fix message links with unbox query
the existing test case didn't include an edge case:

if the unbox key includes a '+' character, when parsed as a query string the '+' becomes a ' ' character, which is no longer a correct unbox key.

this changes the test to match the real-world test case from https://github.com/ssbc/secure-scuttlebutt/pull/220 and fixes the behavior with a hack.