Tidy with ESLint
This codebase has a handful of unused functions, empty block statements,
and unused backslash characters that do nothing. These features add
noise to our syntax and semantics and make the module harder to work on.

This commit applies the recommended ESLint rules, which are a bare
minimum ruleset with only the most conservative best practices.